@@ -1,30 +1,29 @@
 name: flutter_launcher_icons
 description: A package which simplifies the task of updating your Flutter app's launcher icon.
-version: 0.11.0
+version: 0.13.1
 maintainer: Mark O'Sullivan (@MarkOSullivan94)
 homepage: https://github.com/fluttercommunity/flutter_launcher_icons
 repository: https://github.com/fluttercommunity/flutter_launcher_icons/
 issue_tracker: https://github.com/fluttercommunity/flutter_launcher_icons/issues
 
 dependencies:
-  args: ^2.2.0
-  checked_yaml: ^2.0.1
-  cli_util: ^0.3.5
-  image: ^3.0.2
-  json_annotation: ^4.5.0
-  path: ^1.8.0
-  yaml: ^3.1.0
+  args: ^2.3.2
+  checked_yaml: ^2.0.2
+  cli_util: ^0.4.0
+  image: ^4.0.15
+  json_annotation: ^4.8.0
+  path: ^1.8.2
+  yaml: ^3.1.1
 
 environment:
-  sdk: '>=2.13.0 <3.0.0'
+  sdk: '>=2.18.0 <3.0.0'
 
 dev_dependencies:
   # Needed by build_version
-  build_runner: ^2.2.0
-  # allows us to get version number from pubspec yaml which we can pass to Sentry
+  build_runner: ^2.3.3
   # https://pub.dev/packages/build_version
   build_version: ^2.1.1
-  json_serializable: ^6.5.4
+  json_serializable: ^6.6.1
   mockito: ^5.3.2
-  test: ^1.22.0
+  test: ^1.23.1
   test_descriptor: ^2.0.1
