@@ -1,7 +1,8 @@
 import 'dart:io';
 
 import 'package:flutter_launcher_icons/abs/icon_generator.dart';
-import 'package:flutter_launcher_icons/flutter_launcher_icons_config.dart';
+import 'package:flutter_launcher_icons/config/config.dart';
+import 'package:flutter_launcher_icons/config/macos_config.dart';
 import 'package:flutter_launcher_icons/logger.dart';
 import 'package:flutter_launcher_icons/macos/macos_icon_generator.dart';
 import 'package:mockito/annotations.dart';
@@ -13,18 +14,17 @@ import 'package:test_descriptor/test_descriptor.dart' as d;
 import '../templates.dart' as templates;
 
 @GenerateNiceMocks([
-  MockSpec<FlutterLauncherIconsConfig>(),
+  MockSpec<Config>(),
   MockSpec<MacOSConfig>(),
   MockSpec<FLILogger>(),
 ])
-
 import 'macos_icon_generator_test.mocks.dart';
 
 void main() {
   group('MacOSIconGenerator', () {
     late IconGeneratorContext context;
     late IconGenerator generator;
-    late FlutterLauncherIconsConfig mockConfig;
+    late Config mockConfig;
     late MacOSConfig mockMacOSConfig;
     late String prefixPath;
     late File testImageFile;
@@ -38,7 +38,7 @@ void main() {
       });
       setUp(() {
         prefixPath = path.join(d.sandbox, 'fli_test');
-        mockConfig = MockFlutterLauncherIconsConfig();
+        mockConfig = MockConfig();
         mockMacOSConfig = MockMacOSConfig();
         mockLogger = MockFLILogger();
         context = IconGeneratorContext(
@@ -115,7 +115,7 @@ void main() {
   group('MacOSIconGenerator end-to-end', () {
     late IconGeneratorContext context;
     late IconGenerator generator;
-    late FlutterLauncherIconsConfig config;
+    late Config config;
     late String prefixPath;
     final assetPath = path.join(Directory.current.path, 'test', 'assets');
 
@@ -130,7 +130,7 @@ void main() {
         d.file('app_icon.png', imageFile.readAsBytesSync()),
       ]).create();
       prefixPath = path.join(d.sandbox, 'fli_test');
-      config = FlutterLauncherIconsConfig.loadConfigFromPath(
+      config = Config.loadConfigFromPath(
         'flutter_launcher_icons.yaml',
         prefixPath,
       )!;
