@@ -2,7 +2,7 @@ import 'package:test/test.dart';
 
 import 'abs/icon_generator_test.dart' as icon_generator_test;
 import 'android_test.dart' as android_test;
-import 'flutter_launcher_icons_config_test.dart' as fli_config;
+import 'config_test.dart' as fli_config;
 import 'macos/macos_icon_generator_test.dart' as macos_icons_gen_test;
 import 'macos/macos_icon_template_test.dart' as macos_template_test;
 import 'main_test.dart' as main_test;
