@@ -1,5 +1,5 @@
+import 'package:flutter_launcher_icons/config/config.dart';
 import 'package:flutter_launcher_icons/custom_exceptions.dart';
-import 'package:flutter_launcher_icons/flutter_launcher_icons_config.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
@@ -7,7 +7,7 @@ import 'package:test_descriptor/test_descriptor.dart' as d;
 import './templates.dart' as templates;
 
 void main() {
-  group('FlutterLauncherIconsConfig', () {
+  group('Config', () {
     late String prefixPath;
     setUpAll(() {
       prefixPath = path.join(d.sandbox, 'fli_test');
@@ -20,7 +20,7 @@ void main() {
         ]).create();
       });
       test('should return valid configs', () {
-        final configs = FlutterLauncherIconsConfig.loadConfigFromPath(
+        final configs = Config.loadConfigFromPath(
           'flutter_launcher_icons.yaml',
           prefixPath,
         );
@@ -78,7 +78,7 @@ void main() {
       });
 
       test('should return null when invalid filePath is given', () {
-        final configs = FlutterLauncherIconsConfig.loadConfigFromPath(
+        final configs = Config.loadConfigFromPath(
           'file_that_dont_exist.yaml',
           prefixPath,
         );
@@ -87,7 +87,7 @@ void main() {
 
       test('should throw InvalidConfigException when config is invalid', () {
         expect(
-          () => FlutterLauncherIconsConfig.loadConfigFromPath(
+          () => Config.loadConfigFromPath(
             'invalid_fli_config.yaml',
             prefixPath,
           ),
@@ -98,8 +98,7 @@ void main() {
     group('#loadConfigFromTestPubSpec', () {
       test('should return valid configs', () {
         const String path = 'test/config/test_pubspec.yaml';
-        final configs =
-            FlutterLauncherIconsConfig.loadConfigFromPath(path, '.');
+        final configs = Config.loadConfigFromPath(path, '.');
         expect(configs, isNotNull);
         const String imagePath = 'assets/images/icon-710x599.png';
         expect(configs!.imagePath, equals(imagePath));
@@ -132,8 +131,7 @@ void main() {
         ]).create();
       });
       test('should return valid configs', () {
-        final configs =
-            FlutterLauncherIconsConfig.loadConfigFromPubSpec(prefixPath);
+        final configs = Config.loadConfigFromPubSpec(prefixPath);
         expect(configs, isNotNull);
         // android configs
         expect(configs!.android, isTrue);
@@ -200,7 +198,7 @@ void main() {
         });
         test('InvalidConfigException when config is invalid', () {
           expect(
-            () => FlutterLauncherIconsConfig.loadConfigFromPubSpec(prefixPath),
+            () => Config.loadConfigFromPubSpec(prefixPath),
             throwsA(isA<InvalidConfigException>()),
           );
         });
@@ -216,7 +214,7 @@ void main() {
         ]).create();
       });
       test('should return valid config', () {
-        final configs = FlutterLauncherIconsConfig.loadConfigFromFlavor(
+        final configs = Config.loadConfigFromFlavor(
           'development',
           prefixPath,
         );
