@@ -4,9 +4,9 @@
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:cli_util/cli_logging.dart' as _i2;
-import 'package:flutter_launcher_icons/flutter_launcher_icons_config.dart'
-    as _i3;
-import 'package:flutter_launcher_icons/logger.dart' as _i4;
+import 'package:flutter_launcher_icons/config/config.dart' as _i3;
+import 'package:flutter_launcher_icons/config/macos_config.dart' as _i4;
+import 'package:flutter_launcher_icons/logger.dart' as _i5;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -40,11 +40,10 @@ class _FakeProgress_1 extends _i1.SmartFake implements _i2.Progress {
         );
 }
 
-/// A class which mocks [FlutterLauncherIconsConfig].
+/// A class which mocks [Config].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockFlutterLauncherIconsConfig extends _i1.Mock
-    implements _i3.FlutterLauncherIconsConfig {
+class MockConfig extends _i1.Mock implements _i3.Config {
   @override
   int get minSdkAndroid => (super.noSuchMethod(
         Invocation.getter(#minSdkAndroid),
@@ -58,6 +57,12 @@ class MockFlutterLauncherIconsConfig extends _i1.Mock
         returnValueForMissingStub: false,
       ) as bool);
   @override
+  String get backgroundColorIOS => (super.noSuchMethod(
+        Invocation.getter(#backgroundColorIOS),
+        returnValue: '',
+        returnValueForMissingStub: '',
+      ) as String);
+  @override
   bool get hasAndroidAdaptiveConfig => (super.noSuchMethod(
         Invocation.getter(#hasAndroidAdaptiveConfig),
         returnValue: false,
@@ -66,6 +71,24 @@ class MockFlutterLauncherIconsConfig extends _i1.Mock
   @override
   bool get hasPlatformConfig => (super.noSuchMethod(
         Invocation.getter(#hasPlatformConfig),
+        returnValue: false,
+        returnValueForMissingStub: false,
+      ) as bool);
+  @override
+  bool get hasWebConfig => (super.noSuchMethod(
+        Invocation.getter(#hasWebConfig),
+        returnValue: false,
+        returnValueForMissingStub: false,
+      ) as bool);
+  @override
+  bool get hasWindowsConfig => (super.noSuchMethod(
+        Invocation.getter(#hasWindowsConfig),
+        returnValue: false,
+        returnValueForMissingStub: false,
+      ) as bool);
+  @override
+  bool get hasMacOSConfig => (super.noSuchMethod(
+        Invocation.getter(#hasMacOSConfig),
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
@@ -101,7 +124,7 @@ class MockFlutterLauncherIconsConfig extends _i1.Mock
 /// A class which mocks [MacOSConfig].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockMacOSConfig extends _i1.Mock implements _i3.MacOSConfig {
+class MockMacOSConfig extends _i1.Mock implements _i4.MacOSConfig {
   @override
   bool get generate => (super.noSuchMethod(
         Invocation.getter(#generate),
@@ -122,7 +145,7 @@ class MockMacOSConfig extends _i1.Mock implements _i3.MacOSConfig {
 /// A class which mocks [FLILogger].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockFLILogger extends _i1.Mock implements _i4.FLILogger {
+class MockFLILogger extends _i1.Mock implements _i5.FLILogger {
   @override
   bool get isVerbose => (super.noSuchMethod(
         Invocation.getter(#isVerbose),
