@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:args/args.dart';
 import 'package:flutter_launcher_icons/android.dart' as android;
-import 'package:flutter_launcher_icons/flutter_launcher_icons_config.dart';
+import 'package:flutter_launcher_icons/config/config.dart';
 import 'package:flutter_launcher_icons/ios.dart' as ios;
 import 'package:flutter_launcher_icons/main.dart' show defaultConfigFile;
 import 'package:flutter_launcher_icons/main.dart' as main_dart;
@@ -54,29 +54,28 @@ void main() {
     tearDown(() {
       Directory.current = currentDirectory;
     });
+
     test('default', () async {
       await setCurrentDirectory('default');
       await File('flutter_launcher_icons.yaml').writeAsString('''
-flutter_icons:
+flutter_launcher_icons:
   android: true
   ios: false
 ''');
       final ArgResults argResults = parser.parse(<String>[]);
-      final FlutterLauncherIconsConfig? config =
-          main_dart.loadConfigFileFromArgResults(argResults);
+      final Config? config = main_dart.loadConfigFileFromArgResults(argResults);
       expect(config, isNotNull);
       expect(config!.android, isTrue);
     });
     test('default_use_pubspec', () async {
       await setCurrentDirectory('pubspec_only');
       await File('pubspec.yaml').writeAsString('''
-flutter_icons:
+flutter_launcher_icons:
   android: true
   ios: false
 ''');
       ArgResults argResults = parser.parse(<String>[]);
-      final FlutterLauncherIconsConfig? config =
-          main_dart.loadConfigFileFromArgResults(argResults);
+      final Config? config = main_dart.loadConfigFileFromArgResults(argResults);
       expect(config, isNotNull);
       expect(config!.ios, isFalse);
 
@@ -88,14 +87,13 @@ flutter_icons:
     test('custom', () async {
       await setCurrentDirectory('custom');
       await File('custom.yaml').writeAsString('''
-flutter_icons:
+flutter_launcher_icons:
   android: true
   ios: true
 ''');
       // if no argument set, should fail
       ArgResults argResults = parser.parse(<String>['-f', 'custom.yaml']);
-      final FlutterLauncherIconsConfig? config =
-          main_dart.loadConfigFileFromArgResults(argResults);
+      final Config? config = main_dart.loadConfigFileFromArgResults(argResults);
       expect(config, isNotNull);
       expect(config!.ios, isTrue);
 
@@ -115,7 +113,7 @@ flutter_icons:
       'android': true,
       'ios': true
     };
-    final config = FlutterLauncherIconsConfig.fromJson(flutterIconsConfig);
+    final config = Config.fromJson(flutterIconsConfig);
     expect(
       config.getImagePathAndroid(),
       equals('assets/images/icon-710x599.png'),
@@ -126,8 +124,7 @@ flutter_icons:
       'android': true,
       'ios': true
     };
-    final configAndroid =
-        FlutterLauncherIconsConfig.fromJson(flutterIconsConfigAndroid);
+    final configAndroid = Config.fromJson(flutterIconsConfigAndroid);
     expect(
       configAndroid.getImagePathAndroid(),
       equals('assets/images/icon-710x599.png'),
@@ -139,8 +136,7 @@ flutter_icons:
       'android': true,
       'ios': true
     };
-    final configBoth =
-        FlutterLauncherIconsConfig.fromJson(flutterIconsConfigBoth);
+    final configBoth = Config.fromJson(flutterIconsConfigBoth);
     expect(
       configBoth.getImagePathAndroid(),
       equals('assets/images/icon-android.png'),
@@ -154,7 +150,7 @@ flutter_icons:
       'android': true,
       'ios': true
     };
-    final config = FlutterLauncherIconsConfig.fromJson(flutterIconsConfig);
+    final config = Config.fromJson(flutterIconsConfig);
     expect(config.hasPlatformConfig, isTrue);
   });
 
@@ -162,7 +158,7 @@ flutter_icons:
     final Map<String, dynamic> flutterIconsConfig = <String, dynamic>{
       'image_path': 'assets/images/icon-710x599.png'
     };
-    final config = FlutterLauncherIconsConfig.fromJson(flutterIconsConfig);
+    final config = Config.fromJson(flutterIconsConfig);
     expect(config.hasPlatformConfig, isFalse);
   });
 
@@ -172,7 +168,7 @@ flutter_icons:
       'android': false,
       'ios': true
     };
-    final config = FlutterLauncherIconsConfig.fromJson(flutterIconsConfig);
+    final config = Config.fromJson(flutterIconsConfig);
     expect(config.isNeedingNewAndroidIcon, isFalse);
   });
 
@@ -181,7 +177,7 @@ flutter_icons:
       'image_path': 'assets/images/icon-710x599.png',
       'ios': true
     };
-    final config = FlutterLauncherIconsConfig.fromJson(flutterIconsConfig);
+    final config = Config.fromJson(flutterIconsConfig);
     expect(config.isNeedingNewAndroidIcon, isFalse);
   });
 
@@ -191,7 +187,7 @@ flutter_icons:
       'android': true,
       'ios': false
     };
-    final config = FlutterLauncherIconsConfig.fromJson(flutterIconsConfig);
+    final config = Config.fromJson(flutterIconsConfig);
     expect(config.isNeedingNewIOSIcon, isFalse);
   });
 
@@ -200,7 +196,7 @@ flutter_icons:
       'image_path': 'assets/images/icon-710x599.png',
       'android': true
     };
-    final config = FlutterLauncherIconsConfig.fromJson(flutterIconsConfig);
+    final config = Config.fromJson(flutterIconsConfig);
     expect(config.isNeedingNewIOSIcon, isFalse);
   });
 }
